% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_point.R
\name{add_point_to_polygons}
\alias{add_point_to_polygons}
\title{Add lat long to polygons}
\usage{
add_point_to_polygons(map)
}
\arguments{
\item{map}{\code{\link[sp]{SpatialPolygonsDataFrame}} object.}
}
\description{
Each polygon has a labpt point that is approx center of it
This function adds that to SpatialPolygonsDataFrame object
}

