% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/center_coordinate.R
\name{center_coordinate}
\alias{center_coordinate}
\title{This function finds the center of multiple coordinates}
\usage{
center_coordinate(coordinates)
}
\arguments{
\item{coordinates}{list of c(lat, lon) coordinates to be averaged}
}
\description{
This function finds the center of multiple coordinates
}

