% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_gadm.R
\name{clean_gadm}
\alias{clean_gadm}
\title{Cleaning data from gadm}
\usage{
clean_gadm(file_path = system.file("exdata/TUR_adm2.rds", package = "armap"))
}
\arguments{
\item{file_path}{file path of the gadm data to clean. Expected type of data
is explained at \code{inst/exdata/README.md} file.}
}
\description{
This function is for cleaning up the data retrieved from gadm.
}

