% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_spdf.R
\name{convert_spdf}
\alias{convert_spdf}
\title{This function converts a frame with polygons to SpatialPolygonsDataFrame}
\usage{
convert_spdf(frame, polygon_var)
}
\arguments{
\item{frame}{a data frame containing polygon class ojects}

\item{polygon_var}{the column of data frame that contains polygons}
}
\description{
This function converts a frame with polygons to SpatialPolygonsDataFrame
}

