% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_leaflet_box.R
\name{fit_leaflet_box}
\alias{fit_leaflet_box}
\title{This function fits the given leaflet map to polygons bounds}
\usage{
fit_leaflet_box(leaflet, polygons)
}
\arguments{
\item{leaflet}{a leaflet map instance}

\item{polygons}{a polygons object to zoom}
}
\description{
This function fits the given leaflet map to polygons bounds
}

