% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gadm_rename_vars.R
\docType{data}
\name{gadm_city_rename}
\alias{gadm_city_rename}
\title{Renaming city names in gadm data}
\format{An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 24 rows and 2 columns.}
\usage{
gadm_city_rename
}
\description{
City names in map data taken from \href{http://gadm.org}{gadm} does not
align with city names from tuik. This frame is for converting them to the
way they are used in tuik.
}
\keyword{datasets}

