% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gadm.R
\docType{data}
\name{gadm_map}
\alias{gadm_map}
\title{City apartment sales of Turkey}
\format{An object of class \code{SpatialPolygonsDataFrame} with 928 rows and 15 columns.}
\source{
\url{http://www.gadm.org/country} - downloaded for Turkey.
   Check \code{\link{clean_gadm}} function for cleaning of the data.
}
\usage{
gadm_map
}
\description{
A dataset containing maps of Turkey's administrative regions.
Raw data was obtained on September 1, 2016. Check source for details.
}
\keyword{datasets}

