% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gadm_town_parent.R
\docType{data}
\name{gadm_town_parent}
\alias{gadm_town_parent}
\title{Relating gadm towns with tuik towns}
\format{An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 54 rows and 3 columns.}
\usage{
gadm_town_parent
}
\description{
Town maps from \href{http://gadm.org}{gadm} is not complete. Some towns from
there are divided into two or more in reality. This frame describes which
towns that are available from tuik are part of which towns
available in gadm data.
}
\keyword{datasets}

