% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gadm_town_parent.R
\name{gadm_town_parent_self_children}
\alias{gadm_town_parent_self_children}
\title{Town parent data extended with town itself as child}
\usage{
gadm_town_parent_self_children()
}
\description{
\code{\link{gadm_town_parent}} data attaches tuik towns to gadm towns.
However, some of those parents are themselves tuik towns. This function
adds those towns to the \code{gadm_town_parent} as children of themselves.
}

