% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gadm_rename_vars.R
\docType{data}
\name{gadm_town_rename}
\alias{gadm_town_rename}
\title{Renaming town names in gadm data}
\format{An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 35 rows and 3 columns.}
\usage{
gadm_town_rename
}
\description{
Town names in map data taken from \href{http://gadm.org}{gadm} does not
align with town names from tuik. This frame is for converting them to the
way they are used in tuik.
}
\keyword{datasets}

