% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_map.R
\name{get_map}
\alias{get_map}
\title{Map data accessor function}
\usage{
get_map(level = c("town", "city", "country"), istanbul_part = FALSE,
  city_to_region = NULL)
}
\arguments{
\item{level}{granularity of the map.}

\item{istanbul_part}{whether istanbul will be split into two continent parts}

\item{city_to_region}{a frame containing two variables \code{city, region}.
city map will be merged according to those regions}
}
\description{
There is a lot of map data and variations in this package. This function
is for having one point to access them. Through this, maps for towns
cities or the whole country of Turkey can be received. Also, maps polygons
can be clustered as regions.
}

