% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merge.R
\name{merge_region}
\alias{merge_region}
\title{Merge map regions}
\usage{
merge_region(map, matching_region)
}
\arguments{
\item{map}{\code{\link[sp]{SpatialPolygonsDataFrame}}}

\item{matching_region}{matching region for each polygon. Merge will be
done according to these regions}
}
\description{
This function merges polygons in a map given a matching region
}

